% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adatToDGEobj.R
\name{adatToDGEobj}
\alias{adatToDGEobj}
\title{Function adatToDGEobj}
\usage{
adatToDGEobj(adatFile, keepOnlyPasses, sampleNameCol = "ExtIdentifier")
}
\arguments{
\item{adatFile}{Name of a .adat datafile.}

\item{keepOnlyPasses}{TRUE uses Somalogic qc data to exclude failed samples
(Default=TRUE)}

\item{sampleNameCol}{Values of this column should match samples in the
intensity data (Default="ExtIdentifier")}
}
\value{
A DGEobj object with Somalogic data
}
\description{
Reads a Somalogic .adat file and imports the data into a DGEobj.
}
\examples{

 # MySomaObj is class = DGEobj
 MySomaObj <- adatToDGEobj(adatFilename, keepOnlyPasses=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{Somalogic,}
\keyword{adat,}
