% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addItem.R
\name{addItems}
\alias{addItems}
\title{Function addItems (DGEobj)}
\usage{
addItems(dgeObj, itemList, itemTypes, parents, overwrite = FALSE, itemAttr)
}
\arguments{
\item{dgeObj}{A DGEobj that items will be added to. (Required)}

\item{itemList}{A list of data items to add to DGEobj (Required)}

\item{itemTypes}{A list of type values for each item on itemList (Required)}

\item{parents}{A list of parent values for each item on itemList (Optional, but highly recommended)}

\item{overwrite}{Default = FALSE.  Set to TRUE to overwrite the data object
stored in the itemName slot}

\item{itemAttr}{An named list of attributes to add to each item (Optional). The
same set of attributes will be added to each item.}
}
\value{
A class DGEobj object with new items added.
}
\description{
Add a data item to a class DGEobj
}
\examples{
\dontrun{
   # Replace a set of contrasts after adding something to each
   myDGEobj <- addItems(myDGEobj, myContrastList, overwrite= TRUE)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
