% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{appendAttributes}
\alias{appendAttributes}
\title{Function appendAttributes}
\usage{
appendAttributes(dgeObj, itemName, attrNames, values)
}
\arguments{
\item{dgeObj}{a DGEobj structure}

\item{itemName}{Name of the item being annotated}

\item{attrNames}{List of attribute Names to update}

\item{values}{list of Values to append to the named attributes}
}
\value{
the specified attribute value (data type depends on the datatype
stored in the attribute) or NULL if attribute doesn't exist
}
\description{
DGEobj have several attributes whose values are named lists (type, basetype,
parent, funArgs, dateCreated.  The names and length match the names and number
of items.  This
function is used by DGEobj::addItem to update these attributes when a new item
is added.
}
\examples{
   #get an attribute from a DGEobj
   MyDgeObj <- appendAttributes(dgeObj, "counts",
                   list("type", "basetype"),
                   list("counts", "counts"))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

