% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.DGEO.R
\name{as.DGEO}
\alias{as.DGEO}
\title{Function as.DGEO}
\usage{
as.DGEO(RSE)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A DGEobj
}
\description{
Casts an RSE class object as a DGEobj.  Only works with a gene level RSE.
}
\examples{
   MyDGEobj <- as.DGEO(MyRSE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

