% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.DGEobj}
\alias{as.DGEobj}
\title{Function as.DGEobj}
\usage{
as.DGEobj(dgeObj, newClass)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A object of the specified class
}
\description{
Casts a DGEobj class object as the specified data class.
Supports conversion to RangedSummarizedExperiment, Expression Set
or simple list.
}
\examples{
   MyDGElist <- as.list(dgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj,}
\keyword{ExpressionSet}
\keyword{RNA-Seq,}
\keyword{RangedSummarizedExperiment,}

