% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.list.DGEobj}
\alias{as.list.DGEobj}
\title{Function as.list.DGEobj}
\usage{
\method{as.list}{DGEobj}(x, ...)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}

\item{...}{Additional parameters}
}
\value{
A simple list version of the DGEobj.
}
\description{
Casts a DGEobj class object as a simple list.
}
\examples{
\dontrun{
   MyDGElist <- as.list(DGEobj)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
