% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{baseTypes}
\alias{baseTypes}
\title{Function baseTypes}
\usage{
baseTypes(dgeObj)
}
\arguments{
\item{dgeObj}{A class DGEobj object}
}
\value{
A list of baseTypes
}
\description{
Return a list of the available baseTypes in a DGEobj.
}
\examples{
    # Global definition of baseTypes
    baseTypes()

    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    # Basetypes from a specific DGEobj
    baseTypes(exObj)

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
