% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{buildContrastDB}
\alias{buildContrastDB}
\title{Function buildContrastDB}
\usage{
buildContrastDB(inputPath, outputPath, rebuild = FALSE)
}
\arguments{
\item{inputPath}{folder containing DGEobjects saved in .RDS files}

\item{outputPath}{folder for the output database tables}

\item{rebuild}{Force complete rebuilding if TRUE. Default behavior (FALSE)
performs an update, only indexing new .RDS files that have not been previously
indexed.}
}
\value{
TRUE (success) or FALSE (Failure)
}
\description{
Build a constrast database from a set of DGEObjs in a specified input folder.
Write the database files to the output folder.  Optionally, skip files
already indexed.
}
\details{
For addition to the database, DGEobjs should be fully annotated by running
annotateDGEobj first and then saved as .RDS files and placed in the inputPath.

Each time buildContrastDB is run, it skips RDS files that have already
been indexed.  Use rebuild=TRUE to force indexing of all RDS files.
}
\examples{
   success <- buildContrastDB("~/R/RDS", "~/R/contrastDB")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj,}
\keyword{RNA-Seq,}
\keyword{annotation,}
\keyword{attributes}
