% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convertDGEobj}
\alias{convertDGEobj}
\title{Function convertDGEobj}
\usage{
convertDGEobj(dgeObj, Class)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{Class}{One of "RangedSummarizedExperiment", "RSE", "ExpressionSet", "ES"
(case insensitive)}
}
\value{
An object of the specified class
}
\description{
Casts a DGEobj class object as the specified data class.
Supports conversion to RangedSummarizedExperiment, Expression Set,
or simple list.
}
\examples{
\dontrun{
   MyRSE  <- convertDGEobj(DGEobj, "RSE")
   MyES   <- convertDGEobj(DGEobj, "ES")
   MyList <- convertDGEobj(DGEobj, "list")
}

}
\author{
John Thompson
}
\keyword{DGEobj,}
\keyword{ExpressionSet}
\keyword{RNA-Seq,}
\keyword{RangedSummarizedExperiment,}
