% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convertRSE}
\alias{convertRSE}
\title{Function convertRSE}
\usage{
convertRSE(RSE, Class, countsName = "counts")
}
\arguments{
\item{RSE}{An object of class RangedSummarizedExperiment}

\item{Class}{Destination class (one of "DGEobj", "ES", or "ExpressionSet")}

\item{countsName}{If the name of the counts object is not "Counts", use
this argument to tell the function the actual name of the counts assay in the RSE.}
}
\value{
An object of the specified class
}
\description{
Casts a RSE class object as a DGEobj.  Only captures count
data and the associated row and col annotation.
}
\examples{
\dontrun{
   MyDGEobj <- convertRSE(MyRSE, "DGEobj")
   MyES     <- convertRSE(MyRSE, "ES")
}

}
\author{
John Thompson
}
\keyword{DGEobj,}
\keyword{ExpressionSet}
\keyword{RNA-Seq,}
\keyword{SummarizedExperiment,}
