% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEobj.R
\docType{package}
\name{DGEobj}
\alias{DGEobj}
\title{DGEobj}
\description{
This R package is designed to create and use objects of class 
DGEobj, a special version of a list created to contain information about
genes, experiments, and calculations.
}
\section{Function Overview}{


TBD
}

\section{More Information}{

\code{browseVignettes(package = 'DGEobj')}
}

