% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEresult.R
\name{DGEresult}
\alias{DGEresult}
\title{Function DGEresult}
\usage{
DGEresult(data, dataName, dataType)
}
\arguments{
\item{data}{A parcel of data to be added to the DGEresult object.}

\item{dataName}{A unique name for the data parcel}

\item{dataType}{A data type identifier. See .}
}
\value{
A DGEresult object
}
\description{
Initializes an empty DGEresult object.  Optionally, add
the first data element.
}
\examples{
   # init an empty DGEresult object
   myDGEresult <- DGEresult()
   # init and add the first data object.
   myDGEresult <- DGEresult(data = mycounts,
                   dataName = "counts",
                   dataType = "assay")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

