% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.DGEobj}
\alias{dim.DGEobj}
\title{DGEobj dimensions}
\usage{
\method{dim}{DGEobj}(x)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}
}
\value{
An integer vector [r,c] with a length of 2.
}
\description{
Reports the dimensions of the assay slot (row = genes; col = samples) in a DGEobj.
}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
