% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimnames.DGEobj}
\alias{dimnames.DGEobj}
\title{DGEobj dimnames}
\usage{
\method{dimnames}{DGEobj}(x)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}
}
\value{
A list of length 2 containing rownames and colnames of the DGEobj
}
\description{
Returns the rownames and colnames in a list of length 2 for a DGEobj.
}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
