% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Function getAttributes}
\usage{
getAttributes(
  item,
  excludeList = list("dim", "dimnames", "names", "row.names", "class")
)
}
\arguments{
\item{item}{A DGEobj (or any object with attributes)}

\item{excludeList}{A list of attribute names to exclude from the output
(default = list("dim", "dimnames", "names", "row.names"))}
}
\value{
A named list of attribute values for the items
}
\description{
Get all attributes from an item except for any listed in the excludeList argument.
This is intended to capture user-defined attributes by excluding a few standard
attributes like class and dim.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    getAttributes(exObj)

    # Get the formula attribute from the design (if set)
    attr(exObj$design, "formula")

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
