% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItem}
\alias{getItem}
\title{Function getItem}
\usage{
getItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{itemName}{Name of item to retrieve}
}
\value{
The requested data item
}
\description{
Retrieve an item from a DGEobj by item name.
}
\examples{
\dontrun{
   MyCounts <- getItem(DGEobj, "counts")
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
