% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttributes.R
\name{getItemAttributes}
\alias{getItemAttributes}
\title{Function getItemAttributes}
\usage{
getItemAttributes(dgeObj, attrName)
}
\arguments{
\item{dgeObj}{a DGEobj structre}

\item{attrName}{The name of the item attribute to retrieve}
}
\value{
a list of attribute values for the items
}
\description{
get a named attribute from all items
}
\examples{
   #assign attributes to a DGEobj
   MyTypes <- getItemAttributes(dgeObj, "type")
   MyBaseTypes <- getItemAttributes(dgeObj, "basetype")
   MyCreationDates <- getItemAttributes(dgeObj, "dateCreated")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
