% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getType}
\alias{getType}
\title{Function getType}
\usage{
getType(dgeObj, type, parent)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{type}{A single type of list of types to retrieve.  Enter
showTypes(MyDGEobj) to see a list of allowed types.  See addType() function
to define new types.}

\item{parent}{(Optional) Filter return list for common parent (e.g. useful
to select one set of contrast results when multiple fits have been performed)}
}
\value{
A list of requested data items
}
\description{
Retrieve one or more data items from a DGEobj by type.
}
\examples{
\dontrun{
   MyContrastList <- getType(DGEobj, type = "topTable")
   MyRawData      <- getType(DGEobj, type = list("counts", "design", "geneData"))
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
