% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexDGEobj}
\alias{indexDGEobj}
\title{Function indexDGEobj}
\usage{
indexDGEobj(
  dgeObj,
  keys = list("ID", "BMS_PID", "Title", "Description", "Keywords", "Business Unit",
    "Functional Area", "Disease", "Vendor", "PlatformType", "Technology", "LibraryPrep",
    "Organism", "Tissue", "AlignmentReference", "GeneModel", "TBio_Owner", "TA_Owner",
    "LoadData", "ReadLength", "ReadType", "Pipeline")
)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{keys}{A list of attribute keys to retrieve}
}
\value{
A row of DGEobj Project annotation (1 row data.frame)
}
\description{
Retrieve specified attributes from a DGEobj as a named vector. This is intended
to return only user added attributes (i.e. NOT type, basetype, parent) who's
value is class character.  You can also use the keys argument to retrieve
a specific subset of attributes.
}
\examples{
   MyDgeObj <- indexDGEobj(DGEobj, regfile)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj,}
\keyword{RNA-Seq,}
\keyword{annotation,}
\keyword{attributes}
