% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{indexType}
\alias{indexType}
\title{Function indexType}
\usage{
indexType(dgeObj, type, moreAttr)
}
\arguments{
\item{dgeObj}{A a DGEobj class object}

\item{type}{Name of the type to index}

\item{moreAttr}{Additional attributes of the DGEobj to capture as columns
in the returned dataframe. You should only supply attributes that return
character values (not lists or vectors).}
}
\description{
Return a dataframe containing index entries for the items in dgeObj of
the specified type.   Default columns returned are ItemName, ItemType, BaseType,
Parent and DateCreated.  Additional attribute columns will be added as specified
by the moreAttr argument.  Use names(attributes(dgeObj)) to see the attributes
available in the DGEobj.  You should only supply attributes that return
character values (not lists or vectors).
}
\examples{
   # Index topTable contrast data
   MyIndexDF <- indexType(dgeObj, "topTable")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}
