% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDGEobj_Affy.R
\name{initDGEobj_Affy}
\alias{initDGEobj_Affy}
\title{Function initDGEobj_Affy}
\usage{
initDGEobj_Affy(AffyRMA, rowData, colData, level, BrainArrayVer,
  customAttr, allowShortSampleIDs = FALSE, DGEobjDef = .DGEobjDef)
}
\arguments{
\item{AffyRMA}{An RMA intensity matrix in genes (rows) by samples (col) format
with row and column names.}

\item{rowData}{Protein level annotation. rownames must match
rownames in intensities matrix.}

\item{colData}{A dataframe describing the experiment design.  Rownames much match
colnames(intensities).}

\item{customAttr}{An optional (but highly recommended) named list of attributes
to assign to the DGEobj}

\item{allowShortSampleIDs}{Using sequential integers rownames (even if typed as character)
is discouraged and by default will abort the DGEobj creation.  If you have a legitimate
need to have short samplenames composed of numeric characters, you can set this argument to TRUE.
(default = FALSE)}

\item{DGEobjDef}{An object definition. Defaults to the global DGE object definition
(.DGEobjDef) and you usually shouldn't change this unless you're customizing
the object for new data types.}
}
\value{
A DGEobj object with RMA data
}
\description{
Initializes DGEobj with base data (intensities, gene annotation and sample annotation).
This function is NOT typically used directly by end users.  See rXpress::Xpress2DGEO_Affy which
reads data from Xpress and calls this function to build a DGEobj.
}
\examples{

   myDgeObj <- initDGEobj_Affy(intensities = Myintensities,
                            rowData = MyGeneAnnotation,
                            colData = MyDesign,
                            level = "HG-U133_PLUS_2",
                            customAttr = list (PID = "20171025-0001",
                                               XpressID = "12345")
                           )

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}
