% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initDGEobjDef}
\alias{initDGEobjDef}
\title{Instantiate a class DGEobjDef object.}
\usage{
initDGEobjDef(levels, types, uniqueTypes)
}
\arguments{
\item{levels}{A named character vector.  Names represent new
level names.  Values represent the primaryAssay itemName for each level
(optional).  When a level is defined the itemNames and types for the
primary assay data, rowData and colData are automatically established.  The
types argument is only needed to define downstream workflow objects.}

\item{types}{A named character vector of new types where the
values indicate the basetype of a named data type (optional)}

\item{uniqueTypes}{A name or vector of names to add to the
uniqueType list (optional)}
}
\value{
A class DGEobjDef object suitable for use with initDGEobj
}
\description{
Instantiate a class DGEobjDef object.
}
\examples{
    # return the default DGEobj definition
    myDGEobjDef <- initDGEobjDef()

    # Optionally add some new types and levels for metabolomics data
     myDGEobjDef <- initDGEobjDef(
         levels <- "metabolomics",
         types <- c(normalizedIntensity = "assay"))

}
