% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSOMAobj.R
\name{initSOMAobj}
\alias{initSOMAobj}
\title{Function initSOMAobj}
\usage{
initSOMAobj(intensities, rowData, colData, level, customAttr,
  DGEobjDef = .DGEobjDef)
}
\arguments{
\item{intensities}{A intensity matrix in genes (rows) by samples (col) format
with row and column names.}

\item{rowData}{Protein level annotation. rownames must match
rownames in intensities matrix.}

\item{colData}{A dataframe describing the experiment design.  Rownames much match
colnames(intensities).}

\item{customAttr}{An optional (but highly recommended) named list of attributes
to assign to the DGEobj.}

\item{DGEobjDef}{An object definition. Defaults to the global DGE object definition
(.DGEobjDef) and you usually shouldn't change this unless you're customizing
the object for new data types.}
}
\value{
A DGEobj object with Somalogic data
}
\description{
Initializes DGEobj with base data (intensities, gene annotation and sample annotation).
}
\examples{
   # Load and convert Somalogic data
   library(readat)
   soma_adat <- readAdat(adatfilename)
   soma_se <- as.SummarizedExperiment(some_adat)

   myDgeObj <- initSOMAobj(intensities = Myintensities,
                            rowData = MyGeneAnnotation,
                            colData = MyDesign,
                            level = "gene",
                            customAttr = list (PID = "20171025-0001",
                                               XpressID = "12345",
                                               Genome = "Mouse.B38",
                                               GeneModel = "Ensembl.R84")
                           )

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}
