% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inventory}
\alias{inventory}
\title{Function inventory}
\usage{
inventory(dgeObj, verbose = FALSE)
}
\arguments{
\item{dgeObj}{A DGEobj}

\item{verbose}{Add funArgs to the output (Default = FALSE)}
}
\value{
A data.frame summarizing the data contained in the DGEobj
}
\description{
Show the contents of a DGEobj object.  Note if an item is one-dimensional,
the row column reports the length.  This method is also invoked by applying
print() to a DGEobj.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    inventory(exObj)

}
\author{
John Thompson
}
\keyword{RNA-Seq}
