% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemNames.R
\name{itemNames}
\alias{itemNames}
\title{Function itemNames (DGEobj)}
\usage{
itemNames(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A character vector of names
}
\description{
Returns the names of all data items in a DGEobj
}
\examples{
   MyNames <- itemNames(dgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
