% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.DGEobj}
\alias{length.DGEobj}
\title{Function length.DGEobj}
\usage{
\method{length}{DGEobj}(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
integer
}
\description{
Returns the number of data items in a DGEobj.
}
\examples{
   Mylen <- length(dgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

