% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemNames.R
\name{names.DGEobj}
\alias{names.DGEobj}
\title{Function names.DGEobj}
\usage{
\method{names}{DGEobj}(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A character vector of names
}
\description{
Returns the names of all data items in a DGEobj
}
\examples{
   MyNames <- itemNames(dgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

