% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{newType}
\alias{newType}
\title{Function newType}
\usage{
newType(dgeObj, itemType, baseType, uniqueItem = FALSE)
}
\arguments{
\item{dgeObj}{A class DGEobj created by initDGEobj()}

\item{itemType}{The name of the new type to create}

\item{baseType}{The baseType of the new item. One of [row, col, assay, meta]}

\item{uniqueItem}{If set to TRUE, only one instance of the new type is
allowed in a DGEobj}
}
\value{
A DGEobj with a new type definition embedded
}
\description{
Used to customize a DGEobj definition by adding new types.  A baseType
is also declared and whether more than a single instance is allowed.
}
\examples{
\dontrun{
    MyDgeObj <- newType(MyDgeObj,
                        itemType   = "AffyRMA",
                        baseType   = "assay",
                        uniqueItem = TRUE)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
