% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.DGEobj}
\alias{print.DGEobj}
\title{Function print.DGEobj}
\usage{
\method{print}{DGEobj}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{A DGEobj object}

\item{...}{Additional parameters}

\item{verbose}{Add funArgs to the output (Default = FALSE)}
}
\description{
Print a DGEobj object.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    print(exObj)

}
\author{
John Thompson
}
\keyword{RNA-Seq}
