% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DGEresult}
\alias{print.DGEresult}
\title{Function print.DGEresult}
\usage{
\method{print}{DGEresult}(dgeResult, verbose = FALSE, pretty = TRUE, ...)
}
\arguments{
\item{dgeResult}{A DGEresult object.}

\item{verbose}{Add funArgs to the output}

\item{pretty}{prints using knitr::kable if TRUE}
}
\description{
Print a DGEresult object
}
\examples{
  print(myDGEresult)
  print(myDGEresult, verbose=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

