% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItem}
\alias{rmItem}
\title{Function rmItem}
\usage{
rmItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class DGEobj created by initDGEobj()}

\item{itemName}{Name of the item to remove from the DGEobj}
}
\value{
An updated DGEobj
}
\description{
Removes a named data item from a DGEobj.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    exObj <- rmItem(exObj, "design")

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
