% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItems}
\alias{rmItems}
\title{Function rmItems}
\usage{
rmItems(dgeObj, items)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{items}{Either a character vector of item names or a numeric index of items to
remove from the DGEobj Use inventory(DGEobj) to view the indexes of items.}
}
\value{
An updated DGEobj
}
\description{
Removes a vector or list of named data items from a DGEobj.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    exObj <- rmItems(exObj, c("design", "design_orig"))
    exObj <- rmItems(exObj, c(1:2))

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
