% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{setAttributes}
\alias{setAttributes}
\title{Function setAttributes}
\usage{
setAttributes(item, attribs)
}
\arguments{
\item{item}{An object to attach attributes to}

\item{attribs}{A named list of attribute/value pairs}
}
\value{
The item with new attributes added and no existing attributes removed
}
\description{
Set one or more attributes on an object.  You can use this to add attribute
annotation(s) to a DGEobj or to a specific item within a DGEobj.
The setAttributes() function adds the attributes passed to it in
the attribs argument without deleting attributes already present. To remove an attribute,
you can pass NULL as the value. Any named attribute on the attribs argument list
that already exists in the object will be updated.
}
\examples{
\dontrun{
   # Assign attributes to a DGEobj
   MyAttributes <- list(Platform       = "RNA-Seq",
                        Instrument     = "HiSeq",
                        Vendor         = "Unknown",
                        readType       = "PE",
                        readLength     = 75,
                        strandSpecific = TRUE)
   MyDGEobj <- setAttributes(MyDGEobj, MyAttributes)

   # Set attributes on an item inside a DGEobj
   MyAttributes <- list(normalized   = FALSE,
                        LowIntFilter = "FPK >5 in >= 1 group")
   MyDGEObj[["counts"]] <- setAttributes(MyDGEObj[["counts"]], MyAttributes)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
