% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showAttributes}
\alias{showAttributes}
\title{Function showAttributes}
\usage{
showAttributes(
  dgeObj,
  skipList = c("dim", "dimnames", "rownames", "colnames", "listData", "objDef")
)
}
\arguments{
\item{dgeObj}{A DGEobj created by function initDGEobj()}

\item{skipList}{A character vector of attributes to skip. Use this to avoid
printing certain lengthy attributes like rownames.  Defaults to c("dim",
dimnames", "rownames", "colnames", "listData", "objDef")}
}
\value{
Prints a list of attributes and values.
}
\description{
Prints the attributes associated with a DGEobj.  This prints
all attributes regardless of the class of the attribute value.
Use showMeta() if you are only interested in attributes that are
key/value pairs.
}
\examples{
\dontrun{
   showAttributes(myDGEobj)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
