% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showMeta}
\alias{showMeta}
\title{Function showMeta}
\usage{
showMeta(obj)
}
\arguments{
\item{obj}{An object with attributes to examine (often a DGEobj created by function initDGEobj())}
}
\value{
A data frame with the key value pairs from the object's attributes.
}
\description{
Prints the attributes associated with an object with a limit on the length of
the values stored in the attributes.  Use this to examine the key/value metadata
associated with a DGEobj.  Written for use with a DGEobj, but
should function generically on any object with key/value pair attributes.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    showMeta(exObj)

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
