% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{showTypes}
\alias{showTypes}
\title{Function showTypes}
\usage{
showTypes(dgeObj, pretty = TRUE)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{pretty}{TRUE (the default) invokes knitr::kable() to print a
nicely formatted table}
}
\value{
Prints a list of defined "types"
}
\description{
Show the list of all Types defined in the DGEobj.
}
\examples{
\dontrun{
   showTypes(MyDGEobj)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
