% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.DGEobj}
\alias{subset.DGEobj}
\title{Function subset.DGEobj (DGEobj)}
\usage{
\method{subset}{DGEobj}(x, ..., row, col, drop = FALSE, debug = FALSE)
}
\arguments{
\item{x}{A class DGEobj created by initDGEobj()}

\item{...}{Additional parameters}

\item{row}{Row index for the subset}

\item{col}{Col index for the subset}

\item{drop}{Included for compatibility but has no real meaning in the context
of subsetting a DGEobj. So drop = FALSE is the default and changing this
has no effect.}

\item{debug}{Default = FALSE. Set to TRUE to get more information if subsetting a
DGEobj fails with a dimension error.}
}
\value{
A subsetted DGEobj class object
}
\description{
This is the subset.DGEobj function bound to square brackets.
}
\examples{
\dontrun{
   DGEobj <- subset(DGEobj, 1:10, 100:1000)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
