% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summarize}
\alias{summarize}
\title{Function summarize}
\usage{
summarize(dgeObj, verbose = FALSE, ...)
}
\arguments{
\item{dgeObj}{A DGEobj object}

\item{verbose}{Add funArgs to the output (default=FALSE)}
}
\value{
a dataframe summarizing the data contained in the DGEobj
}
\description{
Print a DGEobj object
}
\examples{
  Mydf <- summarize(myDGEobj)
  knitr::kable(summarize(myDGEobj))
  Mydf <- summarize(myDGEobj, verbose=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

