% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textToDGEobj.R
\name{textToDGEobj}
\alias{textToDGEobj}
\title{Function  textToDGEobj}
\usage{
textToDGEobj(
  path,
  counts = "RNA-Seq.Count.Table.txt",
  seqAnnotation = "RNA-Seq.Count.Annotation.txt",
  design = "RNA-Seq.Design.txt",
  level,
  customAttr,
  gz = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{Path for the three data files (required).  For cloud, use s3fs to
mount a S3 bucket to a local folder path.}

\item{counts}{A matrix or dataframe of R genes by C samples (required)
[Default = "RNA-Seq.Count.Table.txt"]}

\item{seqAnnotation}{Gene, isoform or exon level (row) annotation (required)
[Default = "RNA-Seq.Count.Annotation.txt"]}

\item{design}{Sample annotation with expt factors and other sample-associated
data (required) [Default = RNA-Seq.Design.txt"]}

\item{level}{One of "gene", "isoform", "exon" (required)}

\item{customAttr}{A named list of custom attributes to attach to the DGEobj;
Optional but highly encouraged}

\item{gz}{Set to true to automatically add ".gz" to the default file names.  This
is a convenience for when the default files are gzipped.}

\item{verbose}{Output more info while reading data (Default = FALSE)}
}
\value{
A DGEobj
}
\description{
Builds a DGEobj from Omicsoft output files.
}
\details{
The input files may be gzipped, however, in that case you'll need to provide
the filenames in the "counts", "seqAnnotation" and "design" arguments.

Data Format Requirements:
This function is designed for Omicsoft output files:

Count data should have rownames (sequence ids) and column names (sample IDs).
Sequence annotation should also have the same rownames as assays.

If possible, the sequence annotation should include chromosome position data (chr,
start, end, strand).

Sample annotation must have one row for each column in the count table.
rownames(sampleannotation) == colnames(counts).

Function DGEobj::annotateDGEobj reads key=value pairs from a text file to define
attributes.

For sequence annotation: The gene ID (Ensembl ID in our case) is used as the
rowname.  The following fields are also included: "GeneName" (GeneSymbol),
"Transcript." (transcript#), "Strand", "Chromosome", "Start", "End",
"ExonLength (sum of exons)", "Source" (Ensembl Source field), "IsConfounded"
(overlaps another gene).

Certain downstream functionality depends on specific sequence annotation
fields.  For example, Chromosome, Start, End and Strand are required to build
a genomicRanges object.  And many people depend on the source field to filter
for protein-coding genes.  So you can probably build a DGEobj with any old annotation file.
But minimally adhering to these columns of data will insure operability with downstream
tools.
}
\examples{
\dontrun{
    dgeObj <- textToDgeObj(path = "path/to/omicsoft/files", level = "gene", verbose = TRUE)
}
}
\author{
John Thompson
}
\keyword{DGEObj,}
\keyword{Data}
\keyword{Omicsoft,}
\keyword{RNA-Seq,}
\keyword{loading}
