% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateDGEobj}
\alias{annotateDGEobj}
\title{Function annotateDGEobj}
\usage{
annotateDGEobj(dgeObj, annotationFile, keys = NULL)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{annotationFile}{An Omicsoft registration file (in text format, not Excel)}

\item{keys}{A list of keys to look for in the annotationFile and transfer to the DGEobj.
Use keys = NULL to accept all keys in the annotationFile (Default)}
}
\value{
A DGEobj annotated with attributes from the annotation file.
}
\description{
Reads an annotation file  of key/value pairs and attach them attributes to a DGEobj.
The annotation file should be a text file containing key/value pairs separated by an equals
sign.  The keys parameter specifies which key we want to capture as
attributes on the DGEobj. The value will then be the value of that attribute.
}
\examples{
\dontrun{
   MyDgeObj <- annotateDGEobj(DGEobj, annotationFile)
}

}
\author{
John Thompson
}
\keyword{DGEobj,}
\keyword{RNA-Seq,}
\keyword{annotation,}
\keyword{attributes}
