% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateDGEobj}
\alias{annotateDGEobj}
\title{Function annotateDGEobj}
\usage{
annotateDGEobj(dgeObj, regfile, keys = NULL)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{regfile}{An Omicsoft registration file (in text format, not Excel)}

\item{keys}{A list of keys to look for in the regfile and transfer to the DGEobj.
Use keys = NULL to accept all keys in a regfile (Default)}
}
\value{
A DGEobj annotated with attributes.
}
\description{
Read an Omicsoft Registration file and attach attributes to the DGEobj.
}
\details{
Omicsoft registration files contain a [SampleSet] section that contains
metadata about a project.  These are key/value pairs separated by an equals
sign.  The keys parameter specifies which key we want to capture as
attributes on the DGEobj.

This function should work as long as you provide a text file with key/value
pairs separated by equals sign.
}
\examples{
\dontrun{
   MyDgeObj <- annotateDGEobj(DGEobj, regfile)
}

}
\author{
John Thompson
}
\keyword{DGEobj,}
\keyword{RNA-Seq,}
\keyword{annotation,}
\keyword{attributes}
