% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{appendAttributes}
\alias{appendAttributes}
\title{Function appendAttributes}
\usage{
appendAttributes(dgeObj, itemName, attribs)
}
\arguments{
\item{dgeObj}{a DGEobj structure}

\item{itemName}{Name of the item being annotated}

\item{attribs}{A named list of base attributes to annotate itemName}
}
\value{
A dDGEobj with updated attributes
}
\description{
DGEobj have several attributes whose values are named lists (type, basetype,
parent, funArgs, dateCreated.  The names and length match the names and
number of items.  This function is used by DGEobj::addItem to update these
attributes when a new item is added.
}
\details{
This was written to work on DGEobjs but is generalized to work with any object
that can accept attributes.
}
\examples{
   #get an attribute from a DGEobj
   MyDgeObj <- appendAttributes(dgeObj, "counts",
                   list("type", "basetype"),
                   list("counts", "counts"))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
