% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list.DGEobj}
\alias{as.list.DGEobj}
\title{Function as.list.DGEobj}
\usage{
\method{as.list}{DGEobj}(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A simple list version of the DGEobj
}
\description{
Casts a DGEobj class object as a simple list.
}
\examples{
   MyDGElist <- as.list(dgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

