% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.list.DGEobj}
\alias{as.list.DGEobj}
\title{Cast as a simple list}
\usage{
\method{as.list}{DGEobj}(x, ...)
}
\arguments{
\item{x}{A DGEobj}

\item{...}{Additional parameters}
}
\value{
A named list
}
\description{
Cast as a simple list
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("miniObj.RDS", package = "DGEobj"))

    mylist <- as.list(exObj)

}
