% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{baseType}
\alias{baseType}
\title{Function BaseType}
\usage{
baseType(dgeObj, type)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{type}{An item type for which you want the baseType}
}
\value{
A baseType value (character string)
}
\description{
Return the baseType for a given item type in a DGEobj.
}
\examples{
\dontrun{
   MyBaseType <- baseType(dgeObj, type = "DGEList")
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
