% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{baseTypes}
\alias{baseTypes}
\title{Function baseTypes}
\usage{
baseTypes(dgeObj, itemNames)
}
\arguments{
\item{dgeObj}{a class DGEobj object}
}
\value{
A list of  basetypes
}
\description{
Function baseTypes
}
\examples{
   #global definition of baseTypes
   mybasetypes <- baseTypes()
   #basetypes from a specific DGEobj
   mybasetypes <- baseTypes(myDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
\keyword{Return}
\keyword{a}
\keyword{available}
\keyword{basetypes}
\keyword{list}
\keyword{of}
\keyword{the}
