% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{baseTypes}
\alias{baseTypes}
\title{Get a list of the available baseTypes}
\usage{
baseTypes(dgeObj)
}
\arguments{
\item{dgeObj}{(optional) A DGEobj}
}
\value{
A character vector of baseTypes
}
\description{
Get a list of the available baseTypes
}
\examples{
    # Global definition of baseTypes
    baseTypes()

    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    # Basetypes from a specific DGEobj
    baseTypes(exObj)

}
