% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/types.R
\name{baseTypes}
\alias{baseTypes}
\title{Function baseTypes}
\usage{
baseTypes(dgeObj)

baseTypes(dgeObj)
}
\arguments{
\item{dgeObj}{A class DGEobj object}
}
\value{
A list of baseTypes

A list of baseTypes
}
\description{
Return a list of the available baseTypes in a DGEobj.

Return a list of the available baseTypes in a DGEobj.
}
\examples{
\dontrun{
   # Global definition of baseTypes
   myBaseTypes <- baseTypes()

   # Basetypes from a specific DGEobj
   myBaseTypes <- baseTypes(myDGEobj)
}

\dontrun{
   # Global definition of baseTypes
   myBaseTypes <- baseTypes()

   # Basetypes from a specific DGEobj
   myBaseTypes <- baseTypes(myDGEobj)
}

}
\author{
John Thompson

John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
