% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItem.R
\name{baseTypes}
\alias{baseTypes}
\title{Function baseTypes}
\usage{
baseTypes(dgeObj)
}
\arguments{
\item{dgeObj}{A class DGEobj object}
}
\value{
A list of basetypes
}
\description{
Return a list of the available basetypes in a DGEobj.
}
\examples{
\dontrun{
   # Global definition of baseTypes
   myBaseTypes <- baseTypes()

   # Basetypes from a specific DGEobj
   myBaseTypes <- baseTypes(myDGEobj)
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
