% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{convertDGEobj}
\alias{convertDGEobj}
\title{Function convertDGEobj}
\usage{
convertDGEobj(dgeObj, Class)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{Class}{one of "RangedSummarizedExperiment", "RSE", "ExpressionSet", "ES"
(case insensitive)}
}
\value{
A object of the specified class
}
\description{
Casts a DGEobj class object as the specified data class.
Supports conversion to RangedSummarizedExperiment, Expression Set
or simple list.
}
\examples{
   MyRSE <- convertDGEobj(dgeObj, "RSE")
   MyES <- convertDGEobj(dgeObj, "ES")
   MyList <- convertDGEobj(dgeObj, "list")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj,}
\keyword{ExpressionSet}
\keyword{RNA-Seq,}
\keyword{RangedSummarizedExperiment,}
