% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEobj.R
\name{DGEobj}
\alias{DGEobj}
\title{Function DGEobj}
\usage{
DGEobj(data, dataName, dataType)
}
\arguments{
\item{data}{A parcel of data to be added to the DGEresult object.}

\item{dataName}{A unique name for the data parcel}

\item{dataType}{A data type identifier. See .DGEobj$type}
}
\value{
A DGEresult object
}
\description{
Initializes an empty DGEobj.  Optionally, add
the first data element.
}
\examples{
   # init an empty DGEresult object
   myDGEresult <- DGEresult()
   # init and add the first data object.
   myDGEresult <- DGEresult(data = mycounts,
                   dataName = "counts",
                   dataType = "assay")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

