% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim.DGEobj}
\alias{dim.DGEobj}
\title{Function dim}
\usage{
\method{dim}{DGEobj}(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
An integer vector [r,c] with a length of 2.
}
\description{
Reports the dimensions of the assay slot (row = genes; col = samples).
}
\examples{
   dim(MyDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
