% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.DGEobj}
\alias{dim.DGEobj}
\title{Get the "assay" dimensions (row/genes by col/samples)}
\usage{
\method{dim}{DGEobj}(x)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}
}
\value{
An integer vector [r,c] with a length of 2.
}
\description{
Returns the dimensions of the assay data (baseType)
}
