% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{dimnames.DGEobj}
\alias{dimnames.DGEobj}
\title{Function dimnames.DGEobj}
\usage{
\method{dimnames}{DGEobj}(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A list of length 2 containing rownames and colnames.
}
\description{
Returns the rownames and colnames in a list of length 2.
}
\examples{
   MyDims <- dimnames(dgeObj)
   names(MyDims)
   [1] "rownames" "colnames"

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

