% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimnames.DGEobj}
\alias{dimnames.DGEobj}
\title{Get the "assay" names (row/genes by col/samples)}
\usage{
\method{dimnames}{DGEobj}(x)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}
}
\value{
A list of length 2 containing rownames and colnames of the DGEobj
}
\description{
Returns a list of length 2 containing the the assay data names (baseType)
}
