% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttribute}
\alias{getAttribute}
\title{Function getAttribute}
\usage{
getAttribute(item, attrName)
}
\arguments{
\item{item}{a DGEobj or item}

\item{attrName}{Name of the attribute to retrieve.}
}
\value{
the specified attribute value (data type depends on the datatype
stored in the attribute) or NULL if attribute doesn't exist
}
\description{
get a specified attribute from an item
}
\examples{
   #get an attribute from a DGEobj
   MyAttr <- getAttribute(dgeObj, "type")

   #get an attribute from a DGEobj item
   MyAttr <- getAttribute(dgeObj$designMatrix, "formula")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
