% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttribute}
\alias{getAttribute}
\title{Function getAttribute}
\usage{
getAttribute(item, attrName)
}
\arguments{
\item{item}{A DGEobj or item}

\item{attrName}{Name of the attribute to retrieve.}
}
\value{
The specified attribute value (data type depends on the data type
stored in the attribute) or NULL if attribute doesn't exist
}
\description{
Get a specified attribute from an item.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    # Get an attribute from a DGEobj
    getAttribute(exObj, "type")

    # Get an attribute from a DGEobj item
    getAttribute(exObj$designMatrix, "formula")

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
