% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Function getAttributes}
\usage{
getAttributes(item, excludeList = list("dim", "dimnames", "names",
  "row.names", "class"))
}
\arguments{
\item{item}{a DGEobj structre}

\item{excludeList}{A list of attribute names to exclude from the output
(default = list("dim", "dimnames"))}
}
\value{
a list of attribute values for the items
}
\description{
get all attributes from an item
}
\examples{
   #assign attributes to a DGEobj
   MyAttr <- getAttributes(dgeObj$counts)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
