% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{getAttributes}
\alias{getAttributes}
\title{Function getAttributes}
\usage{
getAttributes(
  item,
  excludeList = list("dim", "dimnames", "names", "row.names", "class")
)
}
\arguments{
\item{item}{a DGEobj structure (or any object with attributes)}

\item{excludeList}{A list of attribute names to exclude from the output
(default = list("dim", "dimnames", "names", "row.names"))}
}
\value{
a named list of attribute values for the items
}
\description{
get all attributes from an item except for ones listed on the excludeList.
This is intended to capture user-defined attributes by excluding a few standard
attributes like class and dim.
}
\examples{
   #get attributes from a DGEobj
   MyAttr <- getAttributes(dgeObj)

   #get the formula attribute from the designMatrix
   MyAttr <- attr(dgeObj$designMatrixName, "formula")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
