% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getBaseType}
\alias{getBaseType}
\title{Function getBaseType}
\usage{
getBaseType(dgeObj, baseType)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{baseType}{One or more of: ["row", "col", "assay", "meta"]}
}
\value{
A simple list of data items
}
\description{
Accessor function for DGEobj class objects.  Retrieves all data items of a
given basetype or list of basetypes.
}
\examples{
   Assays <- getBaseType(dgeObj, baseType="assay")
   AssaysAndGeneAnnotation <- getBaseType(degObj, c("assay", "row"))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
