% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getItem.R
\name{getItem}
\alias{getItem}
\title{Function getItem (DGEobj)}
\usage{
getItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{itemName}{An itemName or list of itemNames to retrieve}
}
\value{
A data item or list data items
}
\description{
Casts a DGEobj class object as a simple list.
}
\examples{
   MyCounts <- getItem(dgeObj, "counts")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

