% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItem}
\alias{getItem}
\title{Function getItem}
\usage{
getItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{itemName}{Name of item to retrieve}
}
\value{
A data item
}
\description{
Function getItem
}
\examples{
   MyCounts <- getItem(dgeObj, "counts")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

