% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItems}
\alias{getItems}
\title{Function getItems}
\usage{
getItems(dgeObj, itemNames)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{itemNames}{A list of itemNames to retrieve}
}
\value{
A data item or list data items
}
\description{
Function getItems
}
\examples{
   MyCounts <- getItem(dgeObj, "counts")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

