% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItems}
\alias{getItems}
\title{Retrieve multiple data items by name}
\usage{
getItems(dgeObj, itemNames)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{itemNames}{A character string, character vector, or list names to retrieve}
}
\value{
A list
}
\description{
Retrieve multiple data items by name
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("miniObj.RDS", package = "DGEobj"))

    myList <- getItems(exObj, list("counts", "geneData"))
    names(myList)

}
