% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItems}
\alias{getItems}
\title{Retrieve data items by name}
\usage{
getItems(dgeObj, itemNames)
}
\arguments{
\item{dgeObj}{A DGEobj}

\item{itemNames}{A character string, character vector, or list names to retrieve}
}
\value{
A list
}
\description{
Retrieve data items by name
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    MyCounts <- getItems(exObj, "counts")

}
