% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getItems}
\alias{getItems}
\title{Function getItems}
\usage{
getItems(dgeObj, itemNames)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{itemNames}{A character string, character vector, or list of itemNames to retrieve}
}
\value{
The requested data item or list of data items.
}
\description{
Retrieve items from a DGEobj by item name.
}
\examples{
\dontrun{
   MyCounts <- getItems(DGEobj, "counts")
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
