% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getType}
\alias{getType}
\title{Function getType}
\usage{
getType(dgeObj, type, parent)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{type}{A single type of list of types to retreives.  Enter
showTypes(MyDgeObj) to see a list of allowed types.  See F. addType
to define new types.}

\item{parent}{(Optional) filter return list for common parent (e.g. useful
to select one set of contrast results when multiple fits have been performed)}
}
\value{
A list of data items
}
\description{
Retrieve one of more data items by type.
}
\examples{
   MyContrastList <- getType(dgeObj, type="topTable")
   MyRawData <- getType(dgeObj, type=list("counts", "design", "geneData"))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
