% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEobj.R
\name{initDGEobj}
\alias{initDGEobj}
\title{Function DGEobj}
\usage{
initDGEobj(counts, colDat, rowDat, rowDatType = "geneDat",
  DGEobjDef = .DGEobjDef)
}
\arguments{
\item{counts}{A count matrix or dataframe with row and colnames.}

\item{colDat}{A dataframe describing the experiment design.  Rownames much match
colnames(counts).}

\item{rowDat}{Gene, isoform or exon level annotation. rownames must match
rownames in count matrix.}

\item{DGEobjDef}{An object definition. Defaults to the global DGE object definition
(.DGEobjDef).}
}
\value{
A DGEresult object
}
\description{
Initializes an empty DGEobj.  Optionally, add
the first data element.
}
\examples{
   # init an empty DGEresult object
   myDGEresult <- initDGEobj()
   # init and add the first data object.
   myDGEresult <- DGEresult(data = mycounts,
                   dataName = "counts",
                   dataType = "assay")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

