% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initDGEobjDef}
\alias{initDGEobjDef}
\title{Instantiate a class DGEobjDef object.}
\usage{
initDGEobjDef(types, levels, uniqueTypes)
}
\arguments{
\item{types}{A named character vector of new types where the values indicate the basetype (optional)}

\item{levels}{A name or vector of names for new levels (optional)}

\item{uniqueTypes}{A name or vector of names to add to the uniqueType list (optional)}
}
\value{
A class DGEobjDef object suitable for use with initDGEobj
}
\description{
Instantiate a class DGEobjDef object.
}
\examples{
    # return the default DGEobj definition
    myDGEobjDef <- initDGEobjDef()

    # Optionally add some new types and levels for metabolomics data
    myDGEobjDef <- initDGEobjDef(
                        types <- c(MSQuant = "assay"),
                        levels <- "metabolites",
                        uniqueTypes <- "MSQuant"
                        )


}
