% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{inventory}
\alias{inventory}
\title{Function inventory}
\usage{
inventory(dgeObj, verbose = FALSE)
}
\arguments{
\item{dgeObj}{A DGEobj object}

\item{verbose}{Add funArgs to the output (default=FALSE)}
}
\value{
a dataframe summarizing the data contained in the DGEobj
}
\description{
Show the contents of a DGEobj object.  Note if an item is one dimensional,
the Row column reports the length.  This method is also invoked by aplying
print to a DGEobj.
}
\examples{
  #capture in a dataframe
  Mydf <- inventory(myDGEobj)
  #print a nicely formatted table
  knitr::kable(inventory(myDGEobj))
  #capture dataframe with verbose output
  Mydf <- inventory(myDGEobj, verbose=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}
