% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inventory}
\alias{inventory}
\title{Retrieve the object inventory}
\usage{
inventory(dgeObj, verbose = FALSE)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{verbose}{Include funArgs column in the output (default = FALSE)}
}
\value{
A data.frame summarizing the data contained in the DGEobj
}
\description{
Retrieve the object inventory
}
\examples{
\dontrun{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    inventory(exObj)
}

}
