% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DGEobj}
\alias{print.DGEobj}
\title{Function print.DGEobj}
\usage{
\method{print}{DGEobj}(dgeObj, verbose = FALSE, pretty = TRUE, ...)
}
\arguments{
\item{dgeObj}{A DGEobj object.}

\item{verbose}{Add funArgs to the output}

\item{pretty}{prints using knitr::kable if TRUE}
}
\description{
Print a DGEobj object
}
\examples{
  print(myDGEobj)
  print(myDGEobj, verbose=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}

