% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dgeObj}
\alias{print.dgeObj}
\title{Function print.DGEobj}
\usage{
\method{print}{dgeObj}(dgeObj, digits = NULL, quote = TRUE,
  na.print = NULL, print.gap = NULL, right = FALSE, max = NULL,
  useSource = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{dgeObj}{A DGEobj object}

\item{verbose}{Add funArgs to the output (default=FALSE)}
}
\description{
Print a DGEobj object
}
\examples{
  print(myDGEobj)
  print(myDGEobj, verbose=TRUE)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq}
