% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.DGEobj}
\alias{print.DGEobj}
\title{Print the Inventory}
\usage{
\method{print}{DGEobj}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{A class DGEobj created by function initDGEobj()}

\item{...}{Additional parameters}

\item{verbose}{Include funArgs column in the output (default = FALSE)}
}
\description{
Print the Inventory
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    print(exObj)

}
