% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetDGEobj.R
\name{resetDGEobj}
\alias{resetDGEobj}
\title{Function  resetDGEobj}
\usage{
resetDGEobj(dgeObj, platformType)
}
\arguments{
\item{platformType}{One of "RNA-Seq", "Affymetrix" or "Somalogic".  Only
required if the PlatformType attribute is missing from the DGEobj.}

\item{DGEobj}{A DGEobj that we wish to extract original un-filtered data from.}
}
\value{
A DGEobj containing the original (unfiltered) data.
}
\description{
During a workflow, a DGEobj typically gets filtered down to remove samples
that fail qc or non-expressed genes.  The resetDGEobj function produces a new DGEobj with
the original un-filtered data.
}
\details{
We can restore the original data, because a copy of the original data was stored on the "meta"
layer and the meta layer is not subject to row or column filters.
}
\examples{

  #get back the original data e.g. to start a different analysis
  unfilteredDGEobj <- resetDGEobj(MyHeavilyFilteredDGEobj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{RNA-Seq;}
\keyword{counts;}
\keyword{intensity}
\keyword{low}
