% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItem}
\alias{rmItem}
\title{Function as.list}
\usage{
rmItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{itemName}{}
}
\value{
An updated DGEobj
}
\description{
Removes a named data item from a DGEobj
}
\examples{
   MyDgeObj <- rmItem(MyDgeObj, "design")

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
