% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItem}
\alias{rmItem}
\title{Removes a named data item}
\usage{
rmItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{itemName}{Name of the item to remove}
}
\value{
A DGEobj
}
\description{
Removes a named data item
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("miniObj.RDS", package = "DGEobj"))

    exObj <- rmItem(exObj, "design")

}
