% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R, R/utils.R
\name{rmItem}
\alias{rmItem}
\title{Function rmItem}
\usage{
rmItem(dgeObj, itemName)

rmItem(dgeObj, itemName)
}
\arguments{
\item{dgeObj}{A class DGEobj created by initDGEobj()}

\item{itemName}{Name of the item to remove from the DGEobj}
}
\value{
An updated DGEobj

An updated DGEobj
}
\description{
Removes a named data item from a DGEobj.

Removes a named data item from a DGEobj.
}
\examples{
\dontrun{
   MyDgeObj <- rmItem(MyDgeObj, "design")
}

\dontrun{
   MyDgeObj <- rmItem(MyDgeObj, "design")
}

}
\author{
John Thompson

John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
