% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItems}
\alias{rmItems}
\title{Function rmItems}
\usage{
rmItems(dgeObj, items)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{items}{Either a character vector of item names or a numeric index of items to
remove from the dgeObj.  Use inventory(dgeobj) to view the indexes of items.}
}
\value{
An updated DGEobj
}
\description{
Removes a vector or list of named data items from a DGEobj
}
\examples{
   MyDgeObj <- rmItems(MyDgeObj, c("designMatrix", "designMatrix_Elist"))

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
