% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmItem.R
\name{rmItems}
\alias{rmItems}
\title{Function rmItems}
\usage{
rmItems(dgeObj, items)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}

\item{items}{Either a character vector of item names or a numeric index of items to
remove from the DGEobj Use inventory(DGEobj) to view the indexes of items.}
}
\value{
An updated DGEobj
}
\description{
Removes a vector or list of named data items from a DGEobj.
}
\examples{
\dontrun{
   MyDgeObj <- rmItems(MyDgeObj, c("designMatrix", "designMatrix_Elist"))
}

}
\author{
John Thompson
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
