% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{setAttribute}
\alias{setAttribute}
\title{Function setAttribute}
\usage{
setAttribute(item, attrib, attribName)
}
\arguments{
\item{item}{An object to attach attributes to}

\item{attrib}{An attribute value to add to the item}

\item{attribName}{A name for the attribute}
}
\value{
The item with the new attribute added and no existing attributes removed
}
\description{
Set an attribute on an object.  You can use this to add attribute
annotation(s) to a DGEobj or to a specific item within a DGEobj.
}
\examples{
\dontrun{
   # Assign attribute to a DGEobj
   MyDGEobj <- setAttribute(MyDGEobj, "RNA-SEQ", "Platform")

   # Set attributes on an item inside a DGEobj
   MyDGEObj[["counts"]] <- setAttribute(MyDGEObj[["counts"]], FALSE, "normalized")
}

}
