% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{setAttributes}
\alias{setAttributes}
\title{Function setAttributes}
\usage{
setAttributes(item, attribs)
}
\arguments{
\item{item}{Anything you want to attach attributes to}

\item{attribs}{A named list of attribute/value pairs}
}
\value{
the item with new attributes
}
\description{
Set one or more attributes on an object.  You can use this to add attribute
annotation(s) to a DGEobj or to a specific item within a DGEobj.  Notably,
using the base attributes function to add attributes strips existing
attributes.  The setAttributes function adds the attributes passed to it in
the attribs argument without deleting attributes already present.   However,
A names attribute on the attribs argument list that already exists in the
object will be updated.  The function is generic in that it should work on
other datatypes/classes, not just the DGEobj.
}
\examples{
   #assign attributes to a DGEobj
   MyAttributes <- list(Platform = "RNA-Seq",
           Instrument = "HiSeq",
           Vendor = "BMS",
           readType = "PE",
           readLength = 75,
           strandSpecific = TRUE)
   MyDGEobj <- setAttributes(MyDGEObj, MyAttributes)

   #set attributes on an item inside a DGEobj
   MyAttributes <- list(normalized = FALSE,
                        LowIntFilter = "FPK >5 in >= 1 group"),
   MyDGEObj[["counts"]] <- setAttributes(MyDGEObj[["counts"]], MyAttributes)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
