% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAttributes.R
\name{setAttributes}
\alias{setAttributes}
\title{Function setAttributes}
\usage{
setAttributes(item, attribs)
}
\arguments{
\item{item}{Anything you want to attach attrbutes to}

\item{attribs}{A named list of attribute/value pairs}
}
\value{
the item with new attributes
}
\description{
Set one or more attributes on an object
}
\examples{
   #assign attributes to a DGEobj
   MyAttributes <- list(Platform = "RNA-Seq",
           Instrument = "HiSeq",
           Vendor = "BMS",
           readType = "PE",
           readLength = 75,
           strandSpecific = TRUE)
   MyDGEobj <- setAttributes(MyDGEObj, MyAttributes)

   #set attributes on an item inside a DGEobj
   MyAttributes <- list(normalized = FALSE,
                        LowIntFilter = "FPK >5 in >= 1 group"),
   MyDGEObj[["counts"]] <- setAttributes(MyDGEObj[["counts"]], MyAttributes)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
