% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showAttributes}
\alias{showAttributes}
\title{Function showAttributes}
\usage{
showAttributes(dgeObj)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
Prints a list of attributes and values.
}
\description{
Prints the attributes associated with a DEGobj.
}
\examples{
   showAttributes(MydgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

