% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showAttributes}
\alias{showAttributes}
\title{Function showAttributes}
\usage{
showAttributes(dgeObj, skip)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{skip}{A character vector of attributes to skip. Use this to avoid
printing certain lengthy attributes like rownames.  Defaults to c("dim",
dimnames", "rownames", "colnames", "listData", "objDef")}
}
\value{
Prints a list of attributes and values.
}
\description{
Prints the attributes associated with a DEGobj.
}
\examples{
   showAttributes(MydgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
