% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showAttributes}
\alias{showAttributes}
\title{Print attributes}
\usage{
showAttributes(
  dgeObj,
  skipList = c("dim", "dimnames", "rownames", "colnames", "listData", "objDef")
)
}
\arguments{
\item{dgeObj}{A DGEobj}

\item{skipList}{A character vector of attributes to skip. Use this to avoid
printing certain lengthy attributes like rownames.  Defaults to c("dim",
"dimnames", "rownames", "colnames", "listData", "objDef")}
}
\description{
This function prints all attributes regardless of the class of the attribute value.
}
\details{
*Note* Use showMeta() to only retrieve attributes that are key/value pairs.
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

   showAttributes(exObj)

}
