% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showMeta}
\alias{showMeta}
\title{Function showMeta}
\usage{
showMeta(obj)
}
\arguments{
\item{obj}{A object with attributes to examine created by function initDGEobj}
}
\value{
A data frame with the key value pairs from the object's attributes.
}
\description{
Prints the attributes associated with a object with a limit on the length of
the values stored in the attributes.  Use this to examine the key=value meta
data associated with a DGEobj.  Written for use with DGEobjs but
should function generically on any object with key/value pair attributes.
}
\examples{
   df <- showMeta(MydgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
