% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showMeta}
\alias{showMeta}
\title{Retrieve the Key/Value metadata attributes}
\usage{
showMeta(dgeObj, printed = TRUE)
}
\arguments{
\item{dgeObj}{A DGEobj with attributes}

\item{printed}{Whether to print the list (default = TRUE)}
}
\value{
A data.frame with "Attribute" and "Value" columns
}
\description{
Retrieve the Key/Value metadata attributes
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    showMeta(exObj)

}
