% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{showMeta}
\alias{showMeta}
\title{Retrieve the Key/Value metadata attributes that have a character value and length of 1}
\usage{
showMeta(dgeObj)
}
\arguments{
\item{dgeObj}{A DGEobj with attributes}
}
\value{
A data.frame with "Attribute" and "Value" columns
}
\description{
Retrieve the Key/Value metadata attributes that have a character value and length of 1
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    showMeta(exObj)

}
