% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showTypes.R
\name{showTypes}
\alias{showTypes}
\title{Function as.list}
\usage{
showTypes(dgeObj, pretty = TRUE)
}
\arguments{
\item{dgeObj}{A class dgeObj created by function initDGEobj}

\item{pretty}{TRUE value (the default) invokes knitr::kable to print a
nicely formatted table}
}
\value{
Prints a list of defined "types"
}
\description{
Casts a DGEobj class object as a simple list.
}
\examples{
   showTypes(MyDgeObj)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}

