% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{showTypes}
\alias{showTypes}
\title{Returns and prints the list of all defined types}
\usage{
showTypes(dgeObj)
}
\arguments{
\item{dgeObj}{A class DGEobj created by function initDGEobj()}
}
\value{
data.frame
}
\description{
Returns and prints the list of all defined types
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("miniObj.RDS", package = "DGEobj"))

    showTypes(exObj)

}
