% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{showTypes}
\alias{showTypes}
\title{Returns and prints the list of all defined types}
\usage{
showTypes(dgeObj, printed = TRUE)
}
\arguments{
\item{dgeObj}{A DGEobj}

\item{printed}{Whether to print the list (default = TRUE)}
}
\value{
data.frame
}
\description{
Returns and prints the list of all defined types
}
\examples{
    # example DGEobj
    exObj <- readRDS(system.file("exampleObj.RDS", package = "DGEobj"))

    showTypes(exObj)

}
