% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.DGEobj}
\alias{[.DGEobj}
\title{Subset with square brackets}
\usage{
\method{[}{DGEobj}(x, ...)
}
\arguments{
\item{x}{A DGEobj}

\item{...}{Additional parameters}
}
\value{
A DGEobj
}
\description{
Subset with square brackets
}
