% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.DGEobj}
\alias{[.DGEobj}
\title{Shortcut function for subsetting a DGE object with square brackets.}
\usage{
\method{[}{DGEobj}(x, ...)
}
\arguments{
\item{x}{A class DGEobj created by initDGEobj()}

\item{...}{Additional parameters}
}
\value{
A subsetted DGEobj class object
}
\description{
Shortcut function for subsetting a DGE object with square brackets.
}
