% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.DGEobj}
\alias{subset.DGEobj}
\title{Function subset (DGEobj)}
\usage{
\method{subset}{DGEobj}(DgeObj, row, col, drop = FALSE, debug = FALSE)
}
\arguments{
\item{row}{row index for the subset}

\item{col}{col index for the subset}

\item{drop}{Included for compatibility but has no real meaning in the context
of subsetting a DGEobj.  So drop=FALSE is the default and changing this
has no effect.}

\item{debug}{Default=FALSE.  Set to TRUE to get more information is subsetting a
DGEobj fails with a dimension error.}

\item{dgeObj}{A class dgeObj created by function initDGEobj}
}
\value{
A subsetted DGEobj class object
}
\description{
Add a data item to a class DGEobj
}
\examples{
   DgeObj <- subset(DgeObj, 1:10, 100:1000)

}
\author{
John Thompson, \email{john.thompson@bms.com}
}
\keyword{DGEobj}
\keyword{RNA-Seq,}
